Break up some of the monolithic modules
