+++ base/src/types.rs
@@ -1,0 +1,20 @@
+// =============================================================================
+// MODULES
+// =============================================================================
+
+mod tpm;
+pub use tpm::*;
+mod tpm2b;
+pub use tpm2b::*;
+mod tpma;
+pub use tpma::*;
+mod tpmi;
+pub use tpmi::*;
+mod tpml;
+pub use tpml::*;
+mod tpms;
+pub use tpms::*;
+mod tpmt;
+pub use tpmt::*;
+mod tpmu;
+pub use tpmu::*;
