+++ client-features/Cargo.toml
@@ -6,3 +6,7 @@ edition = "2021"
 [dependencies]
 tpm2-rs-base = { workspace = true }
 tpm2-rs-client = { workspace = true }
+tpm2-rs-errors = { workspace = true }
+
+[dev-dependencies]
+tpm2-rs-marshal = { workspace = true }
