+++ client/Cargo.toml
@@ -9,6 +9,7 @@ tpm-simulator-tests = []
 [dependencies]
 arrayvec = { workspace = true }
 tpm2-rs-base = { workspace = true }
+tpm2-rs-errors = { workspace = true }
 tpm2-rs-marshal = { workspace = true }
 
 [dev-dependencies]
