+++ service/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2021"
 [dependencies]
 hex-literal = { workspace = true }
 tpm2-rs-base = { workspace = true }
+tpm2-rs-errors = { workspace = true }
