+++ service/src/service.rs
@@ -3,8 +3,8 @@ use crate::buffer::{
 };
 use crate::crypto::Crypto;
 use crate::handler::{self, CommandContext, ContextDeps};
-use tpm2_rs_base::constants::Command;
-use tpm2_rs_base::errors::TpmRcError;
+use tpm2_rs_base::types::TPM2CC;
+use tpm2_rs_errors::TpmRcError;
 
 /// Specifies all of the dependent types for `Service`.
 pub trait ServiceDeps {
@@ -80,8 +80,8 @@ impl<'a, Deps: ServiceDeps> Service<'a, Deps> {
             crypto: self.crypto,
         };
 
-        match Command(command_code) {
-            Command::GetRandom => handler::get_random(request, context),
+        match TPM2CC(command_code) {
+            TPM2CC::GetRandom => handler::get_random(request, context),
             _ => Err(TpmRcError::CommandCode),
         }?;
